Extract lsp to separate directory
